mockHpc/requirements.txt
requests==2.23.0
pyyaml==5.3.1
pytest==5.4.1
requests_mock==1.7.0
pytest_httpserver==0.3.4
